chore: cleanup unused code
