+++ apps/www/components/Features/index.tsx
@@ -1,24 +1,12 @@
+import { Badge } from 'ui'
 import Solutions from 'data/Solutions.json'
 import Telemetry from '~/lib/telemetry'
 import gaEvents from '~/lib/gaEvents'
+import SectionContainer from '../Layouts/SectionContainer'
+import ProductIcon from '../ProductIcon'
+import TextLink from '../TextLink'
 import { useTelemetryProps } from 'common/hooks/useTelemetryProps'
 import { useRouter } from 'next/router'
-import Image from 'next/image'
-import SectionContainer from '../Layouts/SectionContainer'
-import { motion } from 'framer-motion'
-import BackedBy from '../BackedBy'
-import { useBreakpoint, useTheme } from 'common'
-import FeatureCard from './FeatureCard'
-import DatabaseVisual from './DatabaseVisual'
-
-const opacityVariant = {
-  default: { opacity: 1, filter: 'grayscale(1)', transition: { duration: 0.1 } },
-  hover: {
-    opacity: 1,
-    filter: 'grayscale(0)',
-    transition: { duration: 0.15 },
-  },
-}
 
 const Features = () => {
   const router = useRouter()
@@ -58,136 +46,41 @@ const Features = () => {
         )
     }
   }
-  const isSm = useBreakpoint(640)
-  const { isDarkMode } = useTheme()
+
+  const IconSections = Object.values(Solutions).map((solution: any) => {
+    const { name, description, icon, label, url } = solution
+
+    return (
+      <div key={name} className="mb-10 space-y-4 md:mb-0">
+        <div className="flex items-center">
+          <ProductIcon icon={icon} />
+          <dt className="text-scale-1200 ml-4 flex flex-row xl:flex-col">{name}</dt>
+        </div>
+
+        <p className="p">{description}</p>
+
+        {label && (
+          <div>
+            <Badge dot>{label}</Badge>
+          </div>
+        )}
+        {url && (
+          <TextLink
+            label={label ? 'Get notified' : 'Learn more'}
+            url={url}
+            onClick={() => sendTelemetryEvent(name)}
+          />
+        )}
+      </div>
+    )
+  })
 
   return (
-    <SectionContainer className="space-y-8 max-w-7xl mt-0 lg:mt-0 !pt-0">
+    <SectionContainer className="space-y-16 pb-0">
       <h3 className="h3 text-center">Build faster and focus on your products</h3>
-      <dl className="grid grid-cols-1 gap-4 sm:grid-cols-2 md:gap-4 lg:gap-6 md:grid-cols-6">
-        <FeatureCard
-          url={Solutions['database'].url}
-          title={Solutions['database'].name}
-          subtitle={
-            <>
-              Every project is a full Postgres database,
-              <br className="inline-block sm:hidden lg:inline-block" /> the world's most trusted
-              relational database.
-            </>
-          }
-          onClick={() => sendTelemetryEvent(name)}
-          image={
-            <motion.div className="absolute inset-0 z-0" variants={opacityVariant}>
-              <DatabaseVisual isDarkMode={isDarkMode} />
-            </motion.div>
-          }
-          classname="col-span-full md:col-span-3"
-        />
-        <FeatureCard
-          url={Solutions['authentication'].url}
-          title={Solutions['authentication'].name}
-          subtitle={
-            <>
-              Add user sign ups and logins,
-              <br className="inline-block sm:hidden lg:inline-block" /> securing your data with Row
-              Level Security.
-            </>
-          }
-          image={
-            <motion.div className="absolute inset-0 z-0" variants={opacityVariant}>
-              <Image
-                src={isDarkMode ? '/images/index/auth-dark.jpg' : '/images/index/auth-light.jpg'}
-                alt="Supabase Authentication feature, hover image with glow"
-                layout="fill"
-                objectPosition="50% 50%"
-                objectFit="cover"
-              />
-            </motion.div>
-          }
-          classname="md:col-span-3 lg:!col-span-2"
-          onClick={() => sendTelemetryEvent(name)}
-        />
-        <FeatureCard
-          url={Solutions['storage'].url}
-          title={Solutions['storage'].name}
-          subtitle={
-            <>
-              Store, organize, and serve large files,
-              <br className="inline-block sm:hidden lg:inline-block" /> from videos to images.
-            </>
-          }
-          image={
-            <motion.div className="absolute inset-0 z-0" variants={opacityVariant}>
-              <Image
-                src={
-                  isDarkMode ? '/images/index/storage-dark.jpg' : '/images/index/storage-light.jpg'
-                }
-                alt="Supabase Storage feature, hover image with glow"
-                layout="fill"
-                objectPosition="50% 50%"
-                objectFit="cover"
-                quality={95}
-              />
-            </motion.div>
-          }
-          classname="md:!col-span-2 md:h-[350px] lg:h-[390px]"
-          onClick={() => sendTelemetryEvent(name)}
-        />
-        <FeatureCard
-          url={Solutions['edge-functions'].url}
-          title={Solutions['edge-functions'].name}
-          subtitle={
-            <>
-              Easily write custom code
-              <br className="inline-block sm:hidden lg:inline-block" /> without deploying or scaling
-              servers.
-            </>
-          }
-          onClick={() => sendTelemetryEvent(name)}
-          image={
-            <motion.div className="absolute inset-0 z-0" variants={opacityVariant}>
-              <Image
-                src={isDarkMode ? '/images/index/edge-dark.jpg' : '/images/index/edge-light.jpg'}
-                alt="Supabase Edge Functions feature, hover image with glow"
-                layout="fill"
-                objectPosition={isSm ? 'bottom' : '50% 50%'}
-                objectFit={isSm ? 'contain' : 'cover'}
-                quality={90}
-              />
-            </motion.div>
-          }
-          classname="md:!col-span-2 md:h-[350px] lg:h-[390px]"
-        />
-        <FeatureCard
-          url={Solutions['realtime'].url}
-          title={Solutions['realtime'].name}
-          subtitle={
-            <>
-              Build multiplayer experiences
-              <br className="inline-block sm:hidden lg:inline-block" /> with realtime data
-              synchronization.
-            </>
-          }
-          image={
-            <motion.div className="absolute inset-0 z-0" variants={opacityVariant}>
-              <Image
-                src={
-                  isDarkMode
-                    ? '/images/index/realtime-dark.jpg'
-                    : '/images/index/realtime-light.jpg'
-                }
-                alt="Supabase Edge Functions feature, hover image with glow"
-                layout="fill"
-                objectPosition="50% 50%"
-                objectFit="cover"
-                quality={95}
-              />
-            </motion.div>
-          }
-          classname="md:!col-span-2 md:h-[350px] lg:h-[390px]"
-        />
+      <dl className="grid grid-cols-1 gap-y-4 sm:grid-cols-2 md:grid-cols-2 md:gap-16 xl:grid-cols-5 lg:gap-x-8 xl:gap-x-12">
+        {IconSections}
       </dl>
-      <BackedBy className="pt-8" layout="horizontal" />
     </SectionContainer>
   )
 }
